MV3 Port: Refactor
